moved the logic from app.py to chat service

+++ convai/utils/config.py
@@ -10,7 +10,7 @@ class Settings(BaseSettings):
     API_TITLE: str = "Conversation AI"
     API_VERSION: str = "1.0.0"
     HOST: str = "0.0.0.0"
-    PORT: float = 8000
+    PORT: int = 8000
     
     # Project Infomation
     PROJECT_ROOT: Path = Path(__file__).parent.parent.parent
